% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_always.R
\name{check_always}
\alias{check_always}
\alias{uncheck_always}
\title{Always check whether R output is really nice}
\usage{
check_always()

uncheck_always()
}
\description{
\code{check_always()} creates a \link[=addTaskCallback]{task callback function}
to check whether all R output is nice.
\code{uncheck_always()} disables the task callback function,
returning your R session to normal behavior.
}
\examples{
check_always()
23 * 3

uncheck_always()
23 * 3
}
\keyword{nice}
