% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_df}
\alias{check_df}
\title{Checking whether any values in a data frame are nice}
\usage{
check_df(data)
}
\arguments{
\item{data}{A data frame (can be a data.frame, tibble, or data.table)}
}
\description{
Checking whether any values in a data frame are nice
}
\examples{
df <- data.frame(a = c(1, 2, 3, 4, 5, 6), b = c(10, 20, 30, 40, 50, 69))
check_df(df)

df <- data.frame(a = c(1, 2, 3), b = c("ten", "twenty", "sixty nine"))
check_df(df)
}
\keyword{nice}
