% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_df}
\alias{check_df}
\title{Checking whether any values in a dataframe are nice}
\usage{
check_df(x)
}
\description{
Checking whether any values in a dataframe are nice
}
\examples{
check_df(69)
}
\keyword{nice}
